+++ runtime/vm/raw_object.h
@@ -2788,6 +2788,8 @@ class ReceivePortLayout : public InstanceLayout {
   VISIT_FROM(ObjectPtr, send_port_)
   SendPortPtr send_port_;
   InstancePtr handler_;
+  StringPtr debug_name_;
+  StackTracePtr allocation_location_;
   VISIT_TO(ObjectPtr, handler_)
 };