+++ pkg/vm_service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 5.4.0
+- Update to version `3.41.0` of the spec.
+- Added `PortList` class.
+- Added `getPorts` RPC.
+- Added optional properties `portId`, `allocationLocation`, and `debugName` to
+  `InstanceRef` and `Instance`.
+
 ## 5.3.0
 - Added support for `dart:io` extensions version 1.4.
 - Update to version `3.40.0` of the spec.